Added tests and refactored.
