+++ ExamTests/TMDBNetworkingTests.swift
@@ -9,62 +9,44 @@ import XCTest
 @testable import Exam
 
 class TMDBNetworkingTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
     
-    func testTMDBfetchMovie_whenGivenRequestWasInCache_NoRequestShouldBeMade() {
-        
-        // TODO: add another TMDB service tests file split between the cached tests and the non cached tests so we can have a properly sut file added to test everything.
+    var sut:TMDBService!
+    var movieFetchedData:Data!
+    var searchedMoviesData:Data!
+    var initiallyFetchedMoviesData:Data!
+
+    override func setUp() {
         
         let urlSessionConfiguration = URLSessionConfiguration.ephemeral
         urlSessionConfiguration.protocolClasses = [MockURLProtocol.self]
         let session = URLSession(configuration: urlSessionConfiguration)
+        
+        sut = TMDBService(urlSession: session)
+        
         let bundle = Bundle(for: TMDBNetworkingTests.self)
-        let url = bundle.url(forResource: "movie_info", withExtension: "json")
-        let data = try! Data(contentsOf: url!)
-        let decodedData = try! TMDBUtilities.jsonDecoder.decode(Movie.self, from: data)
-        let cache = MyCache()
-        let apiKey = "1998bfbd5a042c8835abedd2ca3106d4"
-        let baseAPIURL = "https://api.themoviedb.org/3"
-        let fetchUrl = URL(string: "\(baseAPIURL)/movie/\(338762)")!
-        var queryItems = [URLQueryItem(name:"api_key", value: apiKey)]
-        let params = ["append_to_response" : "videos,credits"]
-        queryItems.append(contentsOf: params.map {URLQueryItem(name: $0.key, value: $0.value)})
-        var urlComponents = URLComponents(url: fetchUrl, resolvingAgainstBaseURL: false)!
-        urlComponents.queryItems  = queryItems
-
-        cache.setObject(decodedData as AnyObject, forKey: urlComponents.url!.absoluteString as AnyObject)
+        let movieInfoUrl = bundle.url(forResource: "movie_info", withExtension: "json")
+        let searchedMoviesUrl = bundle.url(forResource: "search_movies", withExtension: "json")
+        let initiallyFetchedMoviesUrl = bundle.url(forResource: "movie_list", withExtension: "json")
+        movieFetchedData = try! Data(contentsOf: movieInfoUrl!)
+        searchedMoviesData = try! Data(contentsOf: searchedMoviesUrl!)
+        initiallyFetchedMoviesData = try! Data(contentsOf: initiallyFetchedMoviesUrl!)
+    }
         
-        let sut = TMDBService(urlSession: session, cache: cache)
-        let expectation = self.expectation(description: "TMDBService response expectation")
+    override func tearDown() {
         
-        sut.fetchMovie(id: 338762) { result in
-            switch result {
-            case .success(let movie):
-                XCTAssertEqual(movie.id, 338762, "The movies id should be equal but they are not")
-                expectation.fulfill()
-            case .failure(let error):
-                XCTAssertNil(error)
-            }
-        }
-        self.wait(for: [expectation], timeout: 0.1)
+        sut = nil
+        movieFetchedData = nil
+        searchedMoviesData = nil
+        initiallyFetchedMoviesData = nil
+        
     }
     
     func testTMDBfetchMovie_whenGivenInvalidData_shouldReturnError() {
-        let urlSessionConfiguration = URLSessionConfiguration.ephemeral
-        urlSessionConfiguration.protocolClasses = [MockURLProtocol.self]
-        let session = URLSession(configuration: urlSessionConfiguration)
+        
         let stubResponse = HTTPURLResponse(url: URL(string: "aaa")!, statusCode: 210, httpVersion: nil, headerFields: nil)
-        let data = "{\"path\" : \"incorrect\"}".data(using: .utf8)
-        MockURLProtocol.stubResponseData = data
+        let corruptedData = "{\"path\" : \"incorrect\"}".data(using: .utf8)
+        MockURLProtocol.stubResponseData = corruptedData
         MockURLProtocol.stubURLResponse = stubResponse
-        let sut = TMDBService(urlSession: session)
         let expectation = self.expectation(description: "TMDBService response expectation")
         
         sut.fetchMovie(id: 338762) { result in
@@ -81,12 +63,9 @@ class TMDBNetworkingTests: XCTestCase {
     }
     
     func testTMDBfetchMovie_whenGivenAnInvalidStatusCode_shouldReturnError() {
-        let urlSessionConfiguration = URLSessionConfiguration.ephemeral
-        urlSessionConfiguration.protocolClasses = [MockURLProtocol.self]
-        let session = URLSession(configuration: urlSessionConfiguration)
+        
         let stubResponse = HTTPURLResponse(url: URL(string: "aaa")!, statusCode: 410, httpVersion: nil, headerFields: nil)
         MockURLProtocol.stubURLResponse = stubResponse
-        let sut = TMDBService(urlSession: session)
         let expectation = self.expectation(description: "TMDBService response expectation")
         
         sut.fetchMovie(id: 338762) { result in
@@ -102,16 +81,12 @@ class TMDBNetworkingTests: XCTestCase {
     }
     
     func testTMDBfetchMovie_whenGivenASuccessfulResponse_shouldReturnSuccess() {
-        let urlSessionConfiguration = URLSessionConfiguration.ephemeral
-        urlSessionConfiguration.protocolClasses = [MockURLProtocol.self]
-        let session = URLSession(configuration: urlSessionConfiguration)
-        let bundle = Bundle(for: TMDBNetworkingTests.self)
-        let url = bundle.url(forResource: "movie_info", withExtension: "json")
-        let data = try! Data(contentsOf: url!)
+        
         let stubResponse = HTTPURLResponse(url: URL(string: "aaa")!, statusCode: 200, httpVersion: nil, headerFields: nil)
+        
         MockURLProtocol.stubURLResponse = stubResponse
-        MockURLProtocol.stubResponseData = data
-        let sut = TMDBService(urlSession: session)
+        MockURLProtocol.stubResponseData = movieFetchedData
+        
         let expectation = self.expectation(description: "TMDBService response expectation")
         
         sut.fetchMovie(id: 338762) { result in
@@ -123,6 +98,7 @@ class TMDBNetworkingTests: XCTestCase {
                 XCTAssertNil(error)
             }
         }
+        
         self.wait(for: [expectation], timeout: 0.1)
     }
 }
